Fix overflow of min_dtable_update
size_t is 32-bit on 32-bit machines and min_dtable_update was assigned NGHTTP3_QPACK_INT_MAX, which is 64 bits. This failed the tests on 32-bit machines.

There are a few other places in nghttp3_qpack where uint64_t values are assigned to size_t variables and need to be fixed.

+++ lib/includes/nghttp3/nghttp3.h
@@ -1184,8 +1184,8 @@ typedef struct {
 typedef struct {
   uint64_t max_header_list_size;
   uint64_t max_pushes;
-  uint64_t qpack_max_table_capacity;
-  uint64_t qpack_blocked_streams;
+  size_t qpack_max_table_capacity;
+  size_t qpack_blocked_streams;
 } nghttp3_conn_settings;
 
 NGHTTP3_EXTERN void
